third_party/WebKit/LayoutTests/fast/inspector-support/uncaught-dom8-exception-expected.txt

CONSOLE MESSAGE: line 9: Uncaught NotFoundError: Failed to execute 'appendChild' on 'ContainerNode': The new child element is null.
This tests that we are getting the correct message for DOM Exception 8: NOT_FOUND_ERR.
